Move entity loading reducer to its own domain
